rewritten and simplified tests
