+++ test/helpers/utils.js
@@ -10,4 +10,9 @@ utils.emailGenerator = function() {
     return utils.randomNumberGenerator() + '@timekit.io';
 };
 
+utils.tick = function(fn, timing) {
+  if(!timing) { timing = 0; }
+  setTimeout(fn,timing);
+};
+
 module.exports = utils;
